Move tags utilities to inventory module
